Enable setting pool fee charge
