+++ centrifuge-app/src/pages/IssuerPool/index.tsx
@@ -9,7 +9,6 @@ import { IssuerPoolViewLoanTemplatePage } from './Configuration/ViewLoanTemplate
 import { IssuerPoolInvestorsPage } from './Investors'
 import { IssuerPoolLiquidityPage } from './Liquidity'
 import { IssuerPoolOverviewPage } from './Overview'
-import { IssuerPoolFeesPage } from './PoolFees'
 import { IssuerPoolReportingPage } from './Reporting'
 
 export default function IssuerPoolPage() {
@@ -27,7 +26,6 @@ export default function IssuerPoolPage() {
         <Route path={`${path}/assets`} component={IssuerPoolAssetPage} />
         <Route path={`${path}/liquidity`} component={IssuerPoolLiquidityPage} />
         <Route path={`${path}/reporting`} component={IssuerPoolReportingPage} />
-        <Route path={`${path}/pool-fees`} component={IssuerPoolFeesPage} />
         <Route path={path} component={IssuerPoolOverviewPage} />
       </Switch>
       <PoolChangesBanner poolId={poolId} />
