+++ centrifuge-app/src/pages/Pool/index.tsx
@@ -3,6 +3,7 @@ import { Route, Switch, useRouteMatch } from 'react-router'
 import { PoolDetailAssetsTab } from './Assets'
 import { PoolDetailLiquidityTab } from './Liquidity'
 import { PoolDetailOverviewTab } from './Overview'
+import { PoolFeesTab } from './PoolFees'
 import { PoolDetailReportingTab } from './Reporting'
 
 export default function PoolDetailPage() {
@@ -12,6 +13,7 @@ export default function PoolDetailPage() {
       <Route path={`${path}/reporting`} component={PoolDetailReportingTab} />
       <Route path={`${path}/liquidity`} component={PoolDetailLiquidityTab} />
       <Route path={`${path}/assets`} component={PoolDetailAssetsTab} />
+      <Route path={`${path}/fees`} component={PoolFeesTab} />
       <Route path={path} component={PoolDetailOverviewTab} />
     </Switch>
   )
